Checking before install package

+++ scripts/app-system.sh
@@ -42,14 +42,24 @@ declare -a PACKAGES=(
   "zoxide"
 )
 
-echo -e "Remove unused RPM apps\n"
-sudo dnf remove --assumeyes gnome-connections rhythmbox totem gnome-terminal
+status=$(package_status "gnome-connections")
+if [ "$status" -eq 0 ]; then
+  echo -e "Remove unused RPM apps\n"
+  sudo dnf remove --assumeyes gnome-connections rhythmbox totem gnome-terminal
+fi
 
-echo -e "\n\nInstall RPM packages\n"
-sudo dnf install --assumeyes "${PACKAGES[@]}"
+status=$(package_status "btop")
+if [ "$status" -ne 0 ]; then
+  echo -e "\n\nInstall RPM packages\n"
+  sudo dnf install --assumeyes "${PACKAGES[@]}"
+fi
 
-sudo dnf copr enable --assumeyes awood/bat-extras
-sudo dnf install --assumeyes bat-extras
+status=$(package_status "bat-extras")
+if [ "$status" -ne 0 ]; then
+  sudo dnf copr enable --assumeyes awood/bat-extras
+  sudo dnf install --assumeyes bat-extras
+fi
 
 # Set kitty as default system terminal emulator
+[ -e /usr/local/bin/gnome-terminal ] && sudo rm /usr/local/bin/gnome-terminal
 sudo ln -s /usr/bin/kitty /usr/local/bin/gnome-terminal
